Исправили ошибки, отделили grpc часть
