--- .gitignore
@@ -1,2 +1,0 @@
-go.mod
-do.sum
