--- Taskfile.yaml
@@ -1,12 +1,0 @@
-version: "3"  
-
-tasks:  
-  default: 
-    cmds:  
-      - task: generate  
-  generate:  
-    aliases: 
-      - gen  
-    desc: "Generate code from proto files"  
-    cmds: 
-      - protoc -I proto proto/publickey.proto --go_out=./gen/go --go_opt=paths=source_relative --go-grpc_out=./gen/go/ --go-grpc_opt=paths=source_relative
