+++ cmd/main.go
@@ -1,67 +1,78 @@
 package main
 
 import (
-	"UserServiceAuth/internal/app"
 	"UserServiceAuth/internal/config"
-	"UserServiceAuth/lib/logger/handler/slogpretty"
+	"context"
+	"flag"
 	"fmt"
 	"log/slog"
-
+	"net/http"
 	"os"
 	"os/signal"
 	"syscall"
+	"time"
+
+	"UserServiceAuth/internal/httphandler/login"
+	"UserServiceAuth/internal/httphandler/register"
 
 	"github.com/labstack/echo/v4"
 )
 
 func main() {
+	var configPath string
+	flag.StringVar(&configPath, "config", "", "path to config file")
+	flag.Parse()
 
-	cfg := config.MustLoad()
+	// Валидация аргумента "config"
+	if configPath == "" {
+		fmt.Println("Usage: ./yourapp -config <path_to_config_file>")
+		os.Exit(1)
+	}
+
+	cfg := config.MustLoadByPath(configPath)
+
 	fmt.Println(cfg)
 	log := setupLogger(cfg.Env)
-	log.Info("starting app",
-		slog.String("env", cfg.Env),
-		slog.Any("cfg", cfg),
-		slog.Int("port", cfg.GRPC.Port))
-	log.Debug("debage message")
-	log.Error("error message")
-	log.Warn("warn message")
 
-	// Запуск RestApi сервера
+	// Создание сервера Echo
+	e := echo.New()
+	e.POST("/register", register.RegisterHandler)
+	e.POST("/login", login.LognHandler)
+
+	// Запуск сервера в отдельной горутине
 	go func() {
-		s := echo.New()
-		s.POST("/register", RegisterHandler)
-		s.POST("/login", LognHandler)
-		if err := s.Start(":33033"); err != nil {
-			log.Error("failed to start HTTP server")
+		if err := e.Start(cfg.HttpServer.Adress); err != nil && err != http.ErrServerClosed {
+			log.Error("failed to start HTTP server", "error", err)
 		}
 	}()
 
-	// Запуск gRPC сервера
-	application := app.New(log, cfg.GRPC.Port, cfg.StoragePath, cfg.TokenTTL)
-	go application.GRPCSrv.Run()
+	// Ожидание сигнала для остановки сервера
 	stop := make(chan os.Signal, 1)
 	signal.Notify(stop, syscall.SIGTERM, syscall.SIGINT)
-
 	<-stop
 
-	application.GRPCSrv.Stop()
+	log.Info("Shutting down server...")
 
-	log.Info("Gracefully stopped")
+	// Создание контекста с таймаутом для плавной остановки сервера
+	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+
+	if err := e.Shutdown(ctx); err != nil {
+		log.Error("failed to gracefully shutdown the server", "error", err)
+	} else {
+		log.Info("Server gracefully stopped")
+	}
 }
 
 const (
-	envLocal = "local"
-	envDev   = "dev"
-	envProd  = "prod"
+	envDev  = "dev"
+	envProd = "prod"
 )
 
 func setupLogger(env string) *slog.Logger {
 	var log *slog.Logger
 
 	switch env {
-	case envLocal:
-		log = setupPrettySlog()
 	case envDev:
 		log = slog.New(
 			slog.NewJSONHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelDebug}),
@@ -73,16 +84,4 @@ func setupLogger(env string) *slog.Logger {
 	}
 
 	return log
-}
-
-func setupPrettySlog() *slog.Logger {
-	opts := slogpretty.PrettyHandlerOptions{
-		SlogOpts: &slog.HandlerOptions{
-			Level: slog.LevelDebug,
-		},
-	}
-
-	handler := opts.NewPrettyHandler(os.Stdout)
-
-	return slog.New(handler)
 }
