+++ config/local.yaml
@@ -1,12 +1,6 @@
-env: "local" 
+env: "dev" 
 
 storage_path: "./storage/sso.db" 
-
-token_ttl: 2h
-
-grpc:
-  port: 44044
-  timeout: 30s
 
 http_server:
   adress: "localhost:8082"
