+++ go.mod
@@ -2,11 +2,7 @@ module UserServiceAuth
 
 go 1.22.4
 
-require (
-	github.com/stretchr/testify v1.9.0
-	golang.org/x/crypto v0.24.0
-	google.golang.org/grpc v1.64.0
-)
+require github.com/stretchr/testify v1.9.0
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
@@ -16,20 +12,17 @@ require (
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/valyala/bytebufferpool v1.0.0 // indirect
 	github.com/valyala/fasttemplate v1.2.2 // indirect
+	golang.org/x/crypto v0.24.0 // indirect
 )
 
 require (
 	github.com/BurntSushi/toml v1.2.1 // indirect
-	//github.com/Reensef/gRPC-auth-service-protos v0.1.1
-	github.com/fatih/color v1.17.0
 	github.com/ilyakaznacheev/cleanenv v1.5.0
 	github.com/joho/godotenv v1.5.1 // indirect
 	github.com/labstack/echo/v4 v4.12.0
 	golang.org/x/net v0.24.0 // indirect
 	golang.org/x/sys v0.21.0 // indirect
 	golang.org/x/text v0.16.0 // indirect
-	google.golang.org/genproto/googleapis/rpc v0.0.0-20240318140521-94a12d6c2237 // indirect
-	google.golang.org/protobuf v1.33.0
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 	olympos.io/encoding/edn v0.0.0-20201019073823-d3554ca0b0a3 // indirect
 )
