--- internal/app/app.go
@@ -1,26 +1,0 @@
-package app
-
-import (
-	grpcapp "UserServiceAuth/internal/app/grpc"
-	"log/slog"
-	"time"
-)
-
-type App struct {
-	GRPCSrv *grpcapp.App
-}
-
-func New(
-	log *slog.Logger,
-	grpcPort int,
-	storagePath string,
-	tokenTTL time.Duration,
-) *App {
-
-	grpcApp := grpcapp.New(log, grpcPort)
-
-	return &App{
-		GRPCSrv: grpcApp,
-	}
-
-}
