+++ internal/config/config.go
@@ -1,7 +1,6 @@
 package config
 
 import (
-	"flag"
 	"os"
 	"time"
 
@@ -11,13 +10,7 @@ import (
 type Config struct {
 	Env         string           `yaml:"env" env-default:"local"`
 	StoragePath string           `yaml:"storage_path" env-required:"true"`
-	TokenTTL    time.Duration    `yaml:"token_ttl" env-required:"true"`
-	GRPS        GRPSconfig       `yaml:"grpc" env-required:"true"`
 	HttpServer  HttpServerConfig `yaml:"http_server" env-required:"true"`
-}
-type GRPSconfig struct {
-	Port    int           `yaml:"port"`
-	Timeout time.Duration `yaml:"timeout"`
 }
 
 type HttpServerConfig struct {
@@ -26,37 +19,16 @@ type HttpServerConfig struct {
 	Iddle_timeout time.Duration `yaml:"iddle_timeout" env-default:"60s"`
 }
 
-func MustLoad() *Config {
-	configPath := fetchConfigPath()
-	if configPath == "" {
-		panic("config path is empty")
-	}
-	return MustLoadByPath(configPath)
-}
-
 func MustLoadByPath(configPath string) *Config {
-	// check if file exists
+	// Проверка наличия файла
 	if _, err := os.Stat(configPath); os.IsNotExist(err) {
 		panic("config file does not exist: " + configPath)
 	}
+
 	var cfg Config
 	if err := cleanenv.ReadConfig(configPath, &cfg); err != nil {
 		panic("cannot read config: " + err.Error())
 	}
+
 	return &cfg
-}
-
-// fetchConfigPath fetches config path from command line flag or environment variable.
-// Priority: flag > env > default.
-// Default value is empty string.
-func fetchConfigPath() string {
-	var res string
-	flag.StringVar(&res, "config", "", "path to config file")
-	flag.Parse()
-
-	if res == "" {
-		res = os.Getenv("CONFIG_PATH")
-	}
-
-	return res
 }
