--- lib/functions/functions.go
@@ -1,8 +1,0 @@
-package functions
-
-import "golang.org/x/crypto/bcrypt"
-
-func HashPassword(password string) (string, error) {
-	bytes, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
-	return string(bytes), err
-}
