--- proto/publickey.proto
@@ -1,18 +1,0 @@
-syntax = "proto3";
-
-package publickey;
-
-option go_package = "github.com/bledbereq/UserServiceAuth/gen/go;ssov1";
-
-service GetPublicKey {
-  rpc PublicKey (PublicKeyRequest) returns (PublicKeyResponse);
-
-}
-
-// Объект, который отправляется при вызове RPC-метода Register.
-message PublicKeyRequest {
-}
-
-message PublicKeyResponse {
-  string publicKey = 1; 
-}
